Further tuning of PVZP complaint wording
After ~50 complaints have been sent to PVZP in the
first week, it didn't take much time when PVZP started
to strike back. Original template included the word
"registration", which was causing PVZP to ignore the
request by pointing to the fact that they are under no
obligation to register a patient with a general
practitioner. So the complaint template was tuned to remove
this loophole.
Also a possibility to write the complain for yourself
has been added. All dependent-related logic has been
moved under on_behalf_of section.