No more PVZP for kids required starting 1.1.2024
On Jun, 7, 2023 the President of the Czech Republic
has finally signed off on the Novela o Azilu, that
includes the vital change that will solve the 20+ years
ongoing discrimination of non-EU children and finally
let them exercise their right to health. Starting
Jan, 1, 2024 all children holding long-term residence
permits will finally be granted access to public health
insurance that their legal guardians will be paying for
on a monthly basis.
Formerly, the MOI was capping long-term residence permits
for kids by the length of private health insurance contract.
But given current unique circumstances, this capping should
not be done in case parents present a PVZP contract till
the end of 2023. Anyway, it's always better to be safe and
sorry so let's introduce a reminder for the MOI folks, that
PVZP till the end of 2023 is enough.